bump deps + fix config
