+++ package.json
@@ -9,16 +9,16 @@
     "export": "next export"
   },
   "dependencies": {
-    "@mdx-js/loader": "^1.6.22",
-    "@next/mdx": "^12.1.6",
+    "@mdx-js/loader": "^2.1.2",
+    "@next/mdx": "^12.2.2",
     "autoprefixer": "^10.4.7",
-    "next": "^12.1.6",
+    "next": "^12.2.2",
     "postcss": "^8.4.14",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-elm-components": "^1.1.0",
     "react-responsive-carousel": "^3.2.23",
-    "remark-github": "^10.0.1",
+    "remark-github": "^11.2.3",
     "tailwindcss": "^3.1.5"
   }
 }
